refactor/separate-logic-from-layout splits out hooks, logic and views into separate files, adding unit tests as well as the integration tests
